@@ -7,8 +7,8 @@
     "@emotion/react": "^11",
     "@emotion/styled": "^11",
     "@types/node": "^16.7.10",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
+    "@types/react": "^18.0.26",
+    "@types/react-dom": "^18.0.9",
     "@types/react-router-dom": "^5.1.8",
     "@urql/devtools": "^2.0.3",
     "@urql/exchange-auth": "^0.1.6",
@@ -16,8 +16,8 @@
     "@vitejs/plugin-react": "^1.1.3",
     "framer-motion": "^4",
     "graphql": "^15.5.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-icons": "^4.2.0",
     "react-router-dom": "6",
     "typescript": "^4.4.2",
