@@ -11,11 +11,12 @@
     "node": "^16.13.0"
   },
   "dependencies": {
+    "@apollo/server": "^4.3.0",
     "@types/bcryptjs": "^2.4.2",
+    "@types/cors": "^2.8.13",
     "@types/jsonwebtoken": "^8.5.5",
     "@types/node": "16.7.10",
-    "@types/ws": "^7.4.7",
-    "apollo-server-express": "^3.3.0",
+    "@types/ws": "^8.5.3",
     "bcryptjs": "^2.4.3",
     "chalk": "^4.1.2",
     "class-validator": "^0.13.1",
@@ -23,24 +24,25 @@
     "dataloader": "^2.0.0",
     "dotenv": "^10.0.0",
     "express": "^4.17.1",
-    "graphql": "^15.5.2",
-    "graphql-ws": "^5.4.1",
+    "graphql": "^16.6.0",
+    "graphql-ws": "^5.11.2",
     "jsonwebtoken": "^8.5.1",
     "pg": "^8.4.0",
     "reflect-metadata": "^0.1.10",
-    "ts-node": "10.2.1",
-    "type-graphql": "^1.1.1",
-    "typeorm": "0.2.37",
-    "typescript": "4.4.2",
-    "ws": "^8.2.1"
+    "ts-node": "^10.9.1",
+    "type-graphql": "2.0.0-beta.1",
+    "typeorm": "^0.3.11",
+    "typescript": "^4.9.4",
+    "ws": "^8.11.0"
   },
   "scripts": {
     "start": "node dist/index.js",
     "dev": "concurrently -n TSC,SERVER -c blue.bold,yellow.bold \"npx tsc -w\" \"nodemon --ignore client dist/index.js\" ",
-    "typeorm": "ts-node ./node_modules/typeorm/cli.js --config src/ormconfig.ts",
+    "typeorm": "typeorm-ts-node-commonjs",
     "migration:gen": "yarn typeorm migration:generate -n",
-    "migration:run": "yarn typeorm migration:run",
-    "heroku-postbuild": "yarn tsc && yarn --cwd client && yarn --cwd client build"
+    "migration:run": "yarn typeorm migration:run -d src/data-source.ts",
+    "heroku-postbuild": "yarn tsc && yarn --cwd client && yarn --cwd client build",
+    "install:all": "yarn && yarn --cwd client"
   },
   "license": "MIT"
 }
